add generated values

+++ src/main/resources/db/migration/V1_init_tables.sql
@@ -1,6 +1,6 @@
 create table developers
 (
-    id bigint not null primary key,
+    id serial not null primary key,
     firstName varchar(50),
     lastName varchar(50),
     skills varchar(255)
@@ -8,13 +8,13 @@ create table developers
 
 create table teams
 (
-    id bigint not null primary key,
+    id serial not null primary key,
     name varchar(50),
     developers varchar(255)
 );
 
 create table skills
 (
-    id bigint not null primary key,
+    id serial not null primary key,
     name varchar(50)
 );
